refactor: create wallet namespace
